% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{optival_example}
\alias{optival_example}
\title{Locate example datasets shipped with OPTIVAL}
\usage{
optival_example(which = c("example1", "example2", "example2_loadings"))
}
\arguments{
\item{which}{One of "example1", "example2", or "example2_loadings".}
}
\value{
Absolute path to the requested CSV file.
}
\description{
Returns the path to one of the CSV files used in the worked examples 
of Navarro-González, Ferrando, & Morales-Vives (2026).
}
\examples{
\dontrun{
data <- read.csv(optival_example("example1"))
results <- OPTIVAL(as.matrix(data), n_bootstrap = 2000)
}
}
